# src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabros Bot API",
    "version": "0.1.0",
    "description": "Protected alerting and market-analysis APIs. Machine clients use the x-api-key header; the api-key query parameter exists only for legacy compatibility and may leak through logs. The browser admin console can use Firebase ID tokens on documented admin operations when ENABLE_FIREBASE_ADMIN_AUTH=true."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment"
    }
  ],
...
  "paths": {
...
    "/api/news-monitor": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Analyze configured or queried market symbols",
        "operationId": "getNewsMonitor",
        "parameters": [
          {
            "$ref": "#/components/parameters/CryptoSymbols"
          },
          {
            "$ref": "#/components/parameters/StockSymbols"
          },
          {
            "$ref": "#/components/parameters/DryRun"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NewsMonitorAnalysisResult"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "ApiKeyQuery": []
          }
        ]
      },
      "post": {
        "tags": [
          "News"
        ],
        "summary": "Analyze market symbols from a JSON request",
        "operationId": "postNewsMonitor",
        "parameters": [
          {
            "$ref": "#/components/parameters/DryRun"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/NewsMonitor"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/NewsMonitorAnalysisResult"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "ApiKeyQuery": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api-key",
        "description": "Legacy compatibility only. Prefer x-api-key because query strings may be logged."
      },
      "FirebaseBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token for the browser admin console. Requires an admin.viewer or admin.operator custom claim; webhook operations remain API-key-only."
      }
    },
    "parameters": {
...
      "DryRun": {
        "name": "dryRun",
        "in": "query",
        "description": "When true, analyze and return alerts without notification delivery, deduplication cache claims/writes, or signal-outcome persistence.",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
      "CryptoSymbols": {
        "name": "crypto",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "example": "BTCUSDT,ETHUSD"
      },
      "StockSymbols": {
        "name": "stocks",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "example": "NVDA,MSFT"
      },
...
    },
    "requestBodies": {
...
      "NewsMonitor": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NewsMonitorRequest"
            },
            "example": {
              "crypto": [
                "BTCUSDT",
                "ETHUSD"
              ],
              "stocks": [
                "NVDA"
              ],
              "channels": [
                "telegram"
              ],
              "dryRun": false
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Error": {
        "description": "Request or service error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Invalid request",
              "code": "INVALID_REQUEST"
            }
          }
        }
      },
...
      "NewsMonitorAnalysisResult": {
        "description": "News monitor analysis and delivery result",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NewsMonitorResponse"
            },
            "examples": {
              "dryRun": {
                "summary": "Side-effect-free analysis",
                "value": {
                  "success": true,
                  "dryRun": true,
                  "results": [
                    {
                      "symbol": "BTCUSDT",
                      "status": "analyzed",
                      "alert": {
                        "eventCategory": "price_surge",
                        "headline": "Bitcoin breaks resistance",
                        "confidence": 0.85
                      },
                      "deliveryResults": [],
                      "cached": false,
                      "requestId": "req-dryrun123"
                    }
                  ],
                  "summary": {
                    "total": 1,
                    "analyzed": 1,
                    "cached": 0,
                    "timeout": 0,
                    "error": 0,
                    "quota_exhausted": 0,
                    "alerts_sent": 1
                  },
                  "requestedChannels": [
                    "telegram",
                    "whatsapp"
                  ],
                  "deliveredChannels": [],
                  "totalDurationMs": 850,
                  "requestId": "req-dryrun123",
                  "tokenUsage": {}
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "jobId": {
            "type": "string",
            "description": "Durably stored job identifier returned when queue acceptance is indeterminate."
          }
        }
      },
...
      "NewsMonitorRequest": {
        "type": "object",
        "properties": {
          "crypto": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stocks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dryRun": {
            "type": "boolean",
            "default": false,
            "description": "When true, analyze without notification delivery or persistence. The query parameter form is supported for GET and POST."
          },
          "channels": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "telegram",
                "whatsapp",
                "discord"
              ]
            }
          },
          "telegramChatId": {
            "type": "string",
            "minLength": 1
          },
          "whatsappChatId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "NewsMonitorResponse": {
        "type": "object",
        "required": [
          "success",
          "results",
          "summary",
          "requestId"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "dryRun": {
            "type": "boolean",
            "description": "Present and true when analysis generated alerts without dispatching or persisting them."
          },
          "partial_success": {
            "type": "boolean"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JsonObject"
            }
          },
          "summary": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer"
              },
              "analyzed": {
                "type": "integer"
              },
              "cached": {
                "type": "integer"
              },
              "timeout": {
                "type": "integer"
              },
              "error": {
                "type": "integer"
              },
              "quota_exhausted": {
                "type": "integer",
                "description": "Number of symbols that exhausted Gemini 429 RESOURCE_EXHAUSTED retries."
              },
              "alerts_sent": {
                "type": "integer"
              }
            }
          },
          "requestedChannels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Channels selected for delivery; in dry-run mode these are intended channels."
          },
          "deliveredChannels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Channels that actually received alerts; empty in dry-run mode."
          },
          "totalDurationMs": {
            "type": "integer"
          },
          "requestId": {
            "type": "string"
          },
          "tokenUsage": {
            "$ref": "#/components/schemas/JsonObject"
          }
        }
      },
...
    }
  }
}